+++ desktop/cmp/button/AppMenuButton.ts
@@ -11,7 +11,7 @@ import {Icon} from '@xh/hoist/icon';
 import {menu, menuDivider, menuItem, popover} from '@xh/hoist/kit/blueprint';
 import {wait} from '@xh/hoist/promise';
 import {filterConsecutiveMenuSeparators, isOmitted} from '@xh/hoist/utils/impl';
-import {apiDeprecated, withDefault} from '@xh/hoist/utils/js';
+import {withDefault} from '@xh/hoist/utils/js';
 import {clone, isEmpty, isString} from 'lodash';
 import {isValidElement, ReactNode} from 'react';
 
@@ -184,9 +184,7 @@ function parseMenuItems(items: MenuItemLike[]): ReactNode[] {
             if (item === '-') return menuDivider();
             if (!isMenuItem(item)) return item;
 
-            const onClick = item['onClick'];
-            apiDeprecated('AppMenuButton.extraItems.onClick', {test: onClick, msg: 'Use `actionFn` instead', v: 'v56'});
-            const actionFn = item.actionFn ?? onClick;
+            const {actionFn} = item;
 
             // Create menuItem from config
             const cfg = {